nlp_service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
pika==1.3.2